fix : removed trait call in prefetcher plugin
