+++ src/Prefetcher/alias.php
@@ -8,5 +8,4 @@ declare(strict_types=1);
 \class_alias(Narrowspark\Automatic\Prefetcher\Common\AbstractContainer::class, Narrowspark\Automatic\Common\AbstractContainer::class);
 \class_alias(Narrowspark\Automatic\Prefetcher\Common\Contract\Resettable::class, Narrowspark\Automatic\Common\Contract\Resettable::class);
 \class_alias(Narrowspark\Automatic\Prefetcher\Common\Traits\GetGenericPropertyReaderTrait::class, Narrowspark\Automatic\Common\Traits\GetGenericPropertyReaderTrait::class);
-\class_alias(Narrowspark\Automatic\Prefetcher\Common\Traits\GetComposerVersionTrait::class, Narrowspark\Automatic\Common\Traits\GetComposerVersionTrait::class);
 \class_alias(Narrowspark\Automatic\Prefetcher\Common\Contract\Exception\RuntimeException::class, Narrowspark\Automatic\Common\Contract\Exception\RuntimeException::class);
